+++ src/main/java/tech/stackable/hadoop/StackableGroupMapper.java
@@ -23,7 +23,7 @@ public class StackableGroupMapper implements GroupMappingServiceProvider {
   private static final HttpClient HTTP_CLIENT =
       HttpClient.newBuilder().connectTimeout(Duration.ofSeconds(30)).build();
   private final ObjectMapper json;
-  private URI opaUri;
+  private final URI opaUri;
 
   public StackableGroupMapper() {
     // Guaranteed to be only called once (Effective Java: Item 3)
@@ -37,7 +37,7 @@ public class StackableGroupMapper implements GroupMappingServiceProvider {
     try {
       this.opaUri = URI.create(opaMappingUrl);
     } catch (Exception e) {
-      throw new OpaException.UriInvalid(opaUri, e);
+      throw new OpaException.UriInvalid(opaMappingUrl, e);
     }
 
     LOG.debug("OPA mapping URL: {}", opaMappingUrl);